feature(3rd-party-lib): example of using library only in browser
App.module updates will be merged after PR #102

Key is NOT having app.module be the ROOT NgModule, instead (not using
NgModule there at all) just leaving it as a Base place for all Imports /
Components / Directives / Pipes / Providers that can be SHARED between
`app.browser` & `app.server.module`  so that they are the Base/root
NgModules. Otherwise you can't import 3rd party libraries to (as an
example) the browser only, since they need to be on the Root NgModule.